pyproject.toml
[project]
name = "cv-rag-chatbot"
version = "0.1.0"
description = "A RAG chatbot to answer questions about a CV."
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "gradio>=5.34.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
]
license = { text = "MIT" }